Reshuffle the dir structure
