+++ momaland/learning/__init__.py
@@ -1,0 +1,1 @@
+"""Learning module for momaland."""
