+++ momaland/learning/cooperative_momappo/__init__.py
@@ -1,0 +1,1 @@
+"""Cooperative MOMAPPO."""
